Start filling in demo app, make base class
