+++ progress/react/icons/.eslintrc.js
