+++ progress/react/icons/.npmignore
