+++ progress/react/icons/LICENSE
