+++ progress/react/icons/package.json
