+++ progress/react/icons/src/index.ts
@@ -2,3 +2,4 @@ export { Battery } from './Battery';
 export { Heart } from './Heart';
 export { Pie } from './Pie';
 export { Ups } from './Ups';
+export { ProgressIcon } from './ProgressIcon';
