+++ progress/react/icons/src/types.ts
@@ -5,6 +5,10 @@ export type ProgressIconProps = SVGAttributes<SVGElement> & {
     size?: number;
     outlined?: boolean;
     color?: string;
+    showPercentLabel?: boolean;
+    labelPosition?: 'top' | 'bottom' | 'left' | 'right';
+    labelColor?: string;
+    labelSize?: number;
 };
 
 export type PieProgressProps = ProgressIconProps & {
