+++ progress/react/icons/src/utilities.test.ts
