+++ progress/react/icons/src/utilities.ts
